+++ .idea/workspace.xml
@@ -1,8 +1,77 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="PropertiesComponent">
-    <property name="last_opened_file_path" value="C:/DEVEL/Go/GolandProjects" />
-    <property name="nodejs_interpreter_path.stuck_in_default_project" value="undefined stuck path" />
-    <property name="nodejs_npm_path_reset_for_default_project" value="true" />
+  <component name="AutoImportSettings">
+    <option name="autoReloadType" value="ALL" />
+  </component>
+  <component name="ChangeListManager">
+    <list default="true" id="6b4be029-5288-4508-9359-cac2fd3fa3ea" name="Changes" comment="">
+      <change beforePath="$PROJECT_DIR$/.vscode/launch.json" beforeDir="false" afterPath="$PROJECT_DIR$/.vscode/launch.json" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/README.md" beforeDir="false" afterPath="$PROJECT_DIR$/README.md" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/hotelito/.idea/.gitignore" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/hotelito/.idea/cloudbedsnonlambda.iml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/hotelito/.idea/modules.xml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/hotelito/.idea/vcs.xml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmd/hotelito-aws/.aws-sam/build.toml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmd/hotelito-aws/.aws-sam/build/CloudBedsFunction/cloudbeds" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmd/hotelito-aws/.aws-sam/build/template.yaml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmd/hotelito-aws/.aws-sam/cache/6b2979b2-81a8-4a64-9a9e-2cafd3a03e3c/cloudbeds" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmd/hotelito-aws/Makefile" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmd/hotelito-aws/README.TOOLKIT.md" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmd/hotelito-aws/README.md" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmd/hotelito-aws/cloudbeds/cloudbeds" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmd/hotelito-aws/cloudbeds/go.mod" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmd/hotelito-aws/cloudbeds/go.sum" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmd/hotelito-aws/cloudbeds/main.go" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmd/hotelito-aws/cloudbeds/main_test.go" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmd/hotelito-aws/events/event.json" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmd/hotelito-aws/samconfig.toml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmd/hotelito-aws/template.yaml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/go.mod" beforeDir="false" afterPath="$PROJECT_DIR$/go.mod" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/go.sum" beforeDir="false" afterPath="$PROJECT_DIR$/go.sum" afterDir="false" />
+    </list>
+    <option name="SHOW_DIALOG" value="false" />
+    <option name="HIGHLIGHT_CONFLICTS" value="true" />
+    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
+    <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="GOROOT" url="file:///usr/local/go" />
+  <component name="Git.Settings">
+    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
+  </component>
+  <component name="MarkdownSettingsMigration">
+    <option name="stateVersion" value="1" />
+  </component>
+  <component name="ProjectId" id="2RzKfHcrJZwxs5BewklBDguaN7l" />
+  <component name="ProjectViewState">
+    <option name="hideEmptyMiddlePackages" value="true" />
+    <option name="showLibraryContents" value="true" />
+  </component>
+  <component name="PropertiesComponent"><![CDATA[{
+  "keyToString": {
+    "RunOnceActivity.ShowReadmeOnStart": "true",
+    "RunOnceActivity.go.formatter.settings.were.checked": "true",
+    "RunOnceActivity.go.migrated.go.modules.settings": "true",
+    "RunOnceActivity.go.modules.go.list.on.any.changes.was.set": "true",
+    "WebServerToolWindowFactoryState": "false",
+    "go.import.settings.migrated": "true",
+    "go.sdk.automatically.set": "true",
+    "last_opened_file_path": "/Users/xxot/DEVEL_local/cloudbeds-3cx-integration",
+    "node.js.detected.package.eslint": "true",
+    "node.js.selected.package.eslint": "(autodetect)",
+    "nodejs_package_manager_path": "npm",
+    "settings.editor.selected.configurable": "preferences.pluginManager"
+  }
+}]]></component>
+  <component name="RecentsManager">
+    <key name="MoveFile.RECENT_KEYS">
+      <recent name="$PROJECT_DIR$" />
+    </key>
+  </component>
+  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
+  <component name="TypeScriptGeneratedFilesManager">
+    <option name="version" value="3" />
+  </component>
+  <component name="VgoProject">
+    <settings-migrated>true</settings-migrated>
   </component>
 </project>
