+++ cmd/hotelito-aws-lambda/.aws-sam/build.toml
@@ -1,0 +1,13 @@
+# This file is auto generated by SAM CLI build command
+
+[function_build_definitions.ffa5ee0c-7927-49da-9006-5bc2e3f6611f]
+codeuri = "/Users/xxot/DEVEL_local/cloudbeds-3cx-integration/cmd/hotelito-aws-lambda/hotelito"
+runtime = "go1.x"
+architecture = "x86_64"
+handler = "hotelito_lambda"
+source_hash = "1ee57c666224b312d60f1e1bcb7e4e46a332151c59647aef00d58261dab637a4"
+manifest_hash = ""
+packagetype = "Zip"
+functions = ["HotelitoFunction"]
+
+[layer_build_definitions]
