+++ cmd/hotelito-aws-lambda/Makefile
@@ -1,0 +1,4 @@
+.PHONY: build
+
+build:
+	sam build
