--- cmd/hotelito-aws/.aws-sam/build.toml
@@ -1,13 +1,0 @@
-# This file is auto generated by SAM CLI build command
-
-[function_build_definitions.6b2979b2-81a8-4a64-9a9e-2cafd3a03e3c]
-codeuri = "/Users/xxot/DEVEL_local/cloudbeds-3cx-integration/cloudbeds-3cx-go1.x/cloudbeds"
-runtime = "go1.x"
-architecture = "x86_64"
-handler = "cloudbeds"
-source_hash = "04222b843c17b2ae4d5b58416e337bf40d45faa31fc541f842bff660309c5890"
-manifest_hash = ""
-packagetype = "Zip"
-functions = ["CloudBedsFunction"]
-
-[layer_build_definitions]
