--- cmd/hotelito-aws/Makefile
@@ -1,4 +1,0 @@
-.PHONY: build
-
-build:
-	sam build
