--- cmd/hotelito-aws/template.yaml
@@ -1,45 +1,0 @@
-AWSTemplateFormatVersion: '2010-09-09'
-Transform: AWS::Serverless-2016-10-31
-Description: >
-  cloudbeds-3cx-go1.x
-  
-  SAM Template for cloudbeds-3cx-go1.x
-
-# More info about Globals: https://github.com/awslabs/serverless-application-model/blob/master/docs/globals.rst
-Globals:
-  Function:
-    Timeout: 5
-    MemorySize: 128
-
-Resources:
-  CloudBedsFunction:
-    Type: AWS::Serverless::Function # More info about Function Resource: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#awsserverlessfunction
-    Properties:
-      CodeUri: cloudbeds/
-      Handler: cloudbeds
-      Runtime: go1.x
-      Architectures:
-        - x86_64
-      Events:
-        CatchAll:
-          Type: Api # More info about API Event Source: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#api
-          Properties:
-            Path: /hello
-            Method: GET
-      Environment: # More info about Env Vars: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#environment-object
-        Variables:
-          PARAM1: VALUE
-
-Outputs:
-  # ServerlessRestApi is an implicit API created out of Events key under Serverless::Function
-  # Find out more about other implicit resources you can reference within SAM
-  # https://github.com/awslabs/serverless-application-model/blob/master/docs/internals/generated_resources.rst#api
-  CloudbedsAPI:
-    Description: "API Gateway endpoint URL for Prod environment for Cloudbeds-3CX integration Function"
-    Value: !Sub "https://${ServerlessRestApi}.execute-api.${AWS::Region}.amazonaws.com/Prod/hello/"
-  CloudbedsFunction:
-    Description: "Cloudbeds Lambda Function ARN"
-    Value: !GetAtt CloudBedsFunction.Arn
-  CloudbedsFunctionIamRole:
-    Description: "Implicit IAM Role created for Cloudbeds function"
-    Value: !GetAtt CloudBedsFunction.Arn
