+++ go.mod
@@ -3,6 +3,7 @@ module github.com/olegromanchuk/hotelito
 go 1.19
 
 require (
+	github.com/aws/aws-lambda-go v1.41.0
 	github.com/gorilla/mux v1.8.0
 	github.com/joho/godotenv v1.5.1
 	github.com/sirupsen/logrus v1.9.3
